Added script for geNomad and PHASTEST outputs
This script takes in two fasta files, each a concatenation of nucleotide phage regions  identified by geNomad and PHASTEST. It creates a database of the total phage sequence space and BLAST nucleotides for an all-against-all alignment. BLAST hits are paired with their sequences in the database. By further filtering these data, we can arrive at a conservative estimate on which phage regions were equally identified by each program and which were not. 

*Required:
* Concatenated phage genomes from each program.
* Appropriate specification of file locations. 

There are three outputs from this script:
1: combined fasta file of equated and novel sequences.
2: an equivalent CSV file of phage name, sequence, and which program's phage was used in the final dataframe (the one with the longer sequence is chosen).
3: the database of the phage regions across the programs.